Update names and specs for judgments refactor
`resolution/failed-attacks` was *already* returning a map of orders to
their set of judgment vectors. This commit makes the names and specs
reflect that implementation (`failed-attacks` has been renamed
`attack-judgments` for example).

It also makes `attack-judgments` return a map of orders to their
judgment maps (instead of vectors, because the maps are more
meaningful). The test cases are represented as judgment vectors in a
private variable because they're easier to read, and converted to
judgment maps in a public variable.

This commit may do some other small things.