Render game-state to SVG client-side inside of server-side
We want to display orders as the user is inputting them, which makes more sense
to do client-side.

It would have been nice if we could use the XML entities for the positions of
units and supply centers defined in the SVG, but Chrome didn't seem to apply
them to the SVG inserted via javascript (whether done via the SVG DOM or
innerHTML).

I would prefer to use the SVG DOM to insert elements via javascript, but when I
tried that in Chrome the element was added to the document but didn't follow the
<use> tag to render the desired shapes.