Rename support rule to :attacked-from-supported-location
I had a choice of whether to have a `::conflict-situation` spec separate from
`::dt/conflict-rule`, and chose not to since it's not worth the added confusion.
`:attacked-from-supported-location-but-not-dislodged` is the *ideal* name for
the conflict rule, but `:attacked-from-supported-location` is still a reasonable
name, and is *much* clearer when used as a conflict situation in
resolution-iterative, since we have no idea whether it's going to be dislodged
or not at that phase.