Change `:swapped-places-without-convoy` to `:swapped-places`
Before this change, `:swapped-places-without-convoy` was used as the conflict
situation both when a convoy was and wasn't used. Since I don't distinguish
"potential conflict situations" from "rules explaining the resolution of the
conflict", `:swapped-places` will cause less overall confusion. ("these orders
are attempting to swap places" is an example of a potential conflict situation,
while "these orders were in conflict because they attempted to swap places but
neither used a convoy" is an explanation for a resolution).